requirement.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiohttp_socks==0.10.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
bottle==0.13.2
Brotli==1.1.0
browser-cookie3==0.20.1
cachetools==5.5.2
cairocffi==1.7.1
CairoSVG==2.7.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cssselect2==0.8.0
curl_cffi==0.10.0
defusedxml==0.7.1
Deprecated==1.2.18
distro==1.9.0
Django==5.2
django-cors-headers==4.7.0
dotenv==0.9.9
duckduckgo_search==8.0.0
EbookLib==0.18
et_xmlfile==2.0.0
fastapi==0.115.12
Flask==3.1.0
frozenlist==1.5.0
g4f==0.5.0.4
google-auth==2.38.0
google-genai==1.9.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
jeepney==0.9.0
Jinja2==3.1.6
jiter==0.9.0
lxml==5.3.2
lz4==4.4.4
MarkupSafe==3.0.2
mss==10.0.0
multidict==6.2.0
mysql-connector-python==9.2.0
nest-asyncio==1.6.0
nodriver==0.44
numpy==2.2.4
odfpy==1.4.1
openai==1.72.0
openpyxl==3.1.5
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.7
plyer==2.1.0
primp==0.14.0
propcache==0.3.1
proxy_tools==0.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.22.0
pycryptodomex==3.22.0
pydantic==2.11.3
pydantic_core==2.33.1
pyngrok==7.2.3
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-multipart==0.0.20
python-socks==2.7.1
pytz==2025.2
pywebview==5.4
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
sqlparse==0.5.3
starlette==0.46.1
tinycss2==1.4.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.0
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.19.0
